Add task 8,5 update project structure

+++ JAVACORE6/src/main/java/com/andrii/practice/module08/task083/IManageSystem.java
@@ -13,9 +13,9 @@ public interface IManageSystem<T> {
 
     void delete(T obj);
 
-    void deleteById(int id);
+    void deleteById(long id);
 
-    T get(int id);
+    T get(long id);
 
     Double getPrice(T obj);
